+++ RELEASE-NOTES.md
@@ -22,6 +22,7 @@
 
 ## Fixes and improvements
 * Fixed DBT deploy command to properly handle fully qualified names
+* Fixed DBT deploy command to properly handle local directories with dots in names
 
 
 # v3.10.0